+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-09-16
+
+### Added
+
+- Add `<Plug>(Scalpelua)` normal mode mapping
+
+### Fixed
+
+- Replace deprecated `nvim_buf_add_highlight` function
+
+### Changed
+
+- Drop dehighlighting feature
+
 ## [0.2.2] - 2023-06-17
 
 ### Fixed